+++ .eslintrc.js
@@ -1,4 +1,3 @@
-'use strict'
 /*!
 This file is part of CycloneDX JavaScript Library.
 
@@ -35,7 +34,9 @@ module.exports = {
   },
   plugins: [
     /* see https://github.com/lydell/eslint-plugin-simple-import-sort#readme */
-    'simple-import-sort'
+    'simple-import-sort',
+    /* see https://github.com/Stuk/eslint-plugin-header#readme */
+    'header'
   ],
   env: {
     commonjs: true,
@@ -58,9 +59,12 @@ module.exports = {
     {
       files: ['*.js'],
       rules: {
+        // region docs
+        /* see https://github.com/gajus/eslint-plugin-jsdoc */
         'jsdoc/require-param-type': 'error',
         'jsdoc/require-property-type': 'error',
         'jsdoc/require-returns-type': 'error'
+        // region docs
       },
       settings: {
         jsdoc: {
@@ -71,6 +75,15 @@ module.exports = {
     }
   ],
   rules: {
+    // region override rules from plugin 'standard-with-typescript'
+    /* @see https://typescript-eslint.io/rules/consistent-type-imports/ */
+    '@typescript-eslint/consistent-type-imports': ['error', {
+      /* we need our generated declaration files backward compatible to TS3.8
+       * see https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-8.html
+       */
+      fixStyle: 'separate-type-imports'
+    }],
+    // endregion override rules from plugin 'standard-with-typescript'
     // region sort imports/exports
     /* disable other sorters in favour of `simple-import-sort` */
     'import/order': 0,
@@ -101,8 +114,12 @@ module.exports = {
     'jsdoc/require-throws': 'error',
     'jsdoc/require-yields': 0,
     'jsdoc/require-yields-check': 'error',
-    'jsdoc/sort-tags': 'warn'
+    'jsdoc/sort-tags': 'warn',
     // endregion docs
+    // region license-header
+    /* see https://github.com/Stuk/eslint-plugin-header#readme */
+    'header/header': ['error', '.license-header.js']
+    // endregion license-header
   },
   settings: {
     jsdoc: {