+++ src/models/organizationalEntity.ts
@@ -20,7 +20,7 @@ Copyright (c) OWASP Foundation. All Rights Reserved.
 import { Comparable, SortableComparables } from '../_helpers/sortable'
 import { OrganizationalContactRepository } from './organizationalContact'
 
-interface OptionalProperties {
+export interface OptionalOrganizationalEntityProperties {
   name?: OrganizationalEntity['name']
   url?: OrganizationalEntity['url']
   contact?: OrganizationalEntity['contact']
@@ -31,7 +31,7 @@ export class OrganizationalEntity implements Comparable<OrganizationalEntity> {
   url: Set<URL | string>
   contact: OrganizationalContactRepository
 
-  constructor (op: OptionalProperties = {}) {
+  constructor (op: OptionalOrganizationalEntityProperties = {}) {
     this.name = op.name
     this.url = op.url ?? new Set()
     this.contact = op.contact ?? new OrganizationalContactRepository()