+++ src/models/swid.ts
@@ -17,10 +17,11 @@ SPDX-License-Identifier: Apache-2.0
 Copyright (c) OWASP Foundation. All Rights Reserved.
 */
 
-import { isNonNegativeInteger, NonNegativeInteger } from '../types'
-import { Attachment } from './attachment'
+import type { NonNegativeInteger } from '../types'
+import { isNonNegativeInteger } from '../types'
+import type { Attachment } from './attachment'
 
-interface OptionalProperties {
+export interface OptionalSWIDProperties {
   version?: SWID['version']
   patch?: SWID['patch']
   text?: SWID['text']
@@ -43,9 +44,9 @@ export class SWID {
   #tagVersion?: NonNegativeInteger
 
   /**
-   * @throws {TypeError} if {@link op.tagVersion} is neither {@link NonNegativeInteger} nor {@link undefined}
+   * @throws {TypeError} if `op.tagVersion` is neither {@link NonNegativeInteger} nor `undefined`
    */
-  constructor (tagId: SWID['tagId'], name: SWID['name'], op: OptionalProperties = {}) {
+  constructor (tagId: SWID['tagId'], name: SWID['name'], op: OptionalSWIDProperties = {}) {
     this.tagId = tagId
     this.name = name
     this.version = op.version
@@ -60,7 +61,7 @@ export class SWID {
   }
 
   /**
-   * @throws {TypeError} if value is neither {@link NonNegativeInteger} nor {@link undefined}
+   * @throws {TypeError} if value is neither {@link NonNegativeInteger} nor `undefined`
    */
   set tagVersion (value: NonNegativeInteger | undefined) {
     if (value !== undefined && !isNonNegativeInteger(value)) {